% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.StabilizedRegression.R
\name{coef.StabilizedRegression}
\alias{coef.StabilizedRegression}
\title{coefficients function}
\usage{
\method{coef}{StabilizedRegression}(object, predictive_model = FALSE, ...)
}
\arguments{
\item{object}{object of class 'StabilizedRegression'.}

\item{predictive_model}{boolean specifying whether to use the}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Coefficients functions for 'StabilizedRegression' objects.
}
\author{
Niklas Pfister
}
