% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regressor.R
\docType{data}
\name{linear_regressor}
\alias{linear_regressor}
\title{Implementation of linear OLS regressor as R6Class.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
linear_regressor
}
\description{
Implementation of linear OLS regressor as R6Class.
}
\author{
Niklas Pfister
}
\keyword{datasets}
