% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SRanalysis.R
\name{plot.SRanalysis}
\alias{plot.SRanalysis}
\title{plot function}
\usage{
\method{plot}{SRanalysis}(x, x_axis = "SRdiff", varnames = NA, labels = FALSE, ...)
}
\arguments{
\item{x}{object of class 'SRanalysis'.}

\item{x_axis}{either "SRdiff" or "SRpred".}

\item{varnames}{vector of variables names given in same ordering
as columns of X. If NA the variable names saved in the
SRanalysis object are used.}

\item{labels}{boolean specifying whether to print names for all
variables with selection probability greater than 0.5. Only
works if varnames has been specified.}

\item{...}{arguments to be passed to or from other methods.}
}
\description{
Plot functions for 'SRanalysis' objects. Allows to visualize the
stability and predictiveness trade-off of individual predictors.
}
\author{
Niklas Pfister
}
