% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.StabilizedRegression.R
\name{predict.StabilizedRegression}
\alias{predict.StabilizedRegression}
\title{predict function}
\usage{
\method{predict}{StabilizedRegression}(object, newdata, predictive_model = FALSE, ...)
}
\arguments{
\item{object}{object of class 'StabilizedRegression'.}

\item{newdata}{matrix or data.frame for which the response should
be predicted.}

\item{predictive_model}{boolean. If \code{TRUE} the model SR
(pred) is used to predict, if \code{FALSE} the model SR is used.}

\item{...}{additional arguments affecting the prediction produced.}
}
\description{
Predict functions for 'StabilizedRegression' objects.
}
\author{
Niklas Pfister
}
